.dev/mcp/requirements.txt
mcp>=1.26.0
httpx>=0.27
PyYAML>=6.0
python-dotenv>=1.0
openai>=1.30
anthropic>=0.34
